requirements.txt
litellm==1.50.0
python-dotenv==1.0.0
requests==2.31.0
duckduckgo-search==6.1.0
matplotlib>=3.8.0
plotly>=5.18.0
pandas>=2.1.4
pillow>=10.0.0